+++ packages/contracts/test/voting/erc20-voting-setup.ts
@@ -17,15 +17,17 @@ const abiCoder = ethers.utils.defaultAbiCoder;
 const AddressZero = ethers.constants.AddressZero;
 const EMPTY_DATA = '0x';
 
+const prepareInstallDataTypes = [
+  'uint64',
+  'uint64',
+  'uint64',
+  'tuple(address,string,string)',
+  'tuple(address[],uint256[])',
+];
+
 // minimum bytes for `prepareInstallation` data param.
 const MINIMUM_DATA = abiCoder.encode(
-  [
-    'uint64',
-    'uint64',
-    'uint64',
-    'tuple(address,string,string)',
-    'tuple(address[],uint256[])',
-  ],
+  prepareInstallDataTypes,
   [1, 1, 1, [AddressZero, '', ''], [[], []]]
 );
 
@@ -44,7 +46,6 @@ const SET_CONFIGURATION_PERMISSION_ID = ethers.utils.id(
 const UPGRADE_PERMISSION_ID = ethers.utils.id('UPGRADE_PLUGIN_PERMISSION');
 const EXECUTE_PERMISSION_ID = ethers.utils.id('EXECUTE_PERMISSION');
 const MINT_PERMISSION_ID = ethers.utils.id('MINT_PERMISSION');
-const MERKLE_MINT_PERMISSION_ID = ethers.utils.id('MERKLE_MINT_PERMISSION');
 
 describe('ERC20VotingSetup', function () {
   let ownerAddress: string;
@@ -114,13 +115,7 @@ describe('ERC20VotingSetup', function () {
 
     it('fails if `MintSettings` arrays do not have the same length', async () => {
       const data = abiCoder.encode(
-        [
-          'uint64',
-          'uint64',
-          'uint64',
-          'tuple(address,string,string)',
-          'tuple(address[],uint256[])',
-        ],
+        prepareInstallDataTypes,
         [1, 1, 1, [AddressZero, '', ''], [[AddressZero], []]]
       );
 
@@ -132,13 +127,7 @@ describe('ERC20VotingSetup', function () {
     it('fails if passed token address is not a contract', async () => {
       const tokenAddress = ownerAddress;
       const data = abiCoder.encode(
-        [
-          'uint64',
-          'uint64',
-          'uint64',
-          'tuple(address,string,string)',
-          'tuple(address[],uint256[])',
-        ],
+        prepareInstallDataTypes,
         [1, 1, 1, [tokenAddress, '', ''], [[], []]]
       );
 
@@ -150,13 +139,7 @@ describe('ERC20VotingSetup', function () {
     it('fails if passed token address is not ERC20', async () => {
       const tokenAddress = implementationAddress;
       const data = abiCoder.encode(
-        [
-          'uint64',
-          'uint64',
-          'uint64',
-          'tuple(address,string,string)',
-          'tuple(address[],uint256[])',
-        ],
+        prepareInstallDataTypes,
         [1, 1, 1, [tokenAddress, '', ''], [[], []]]
       );
 
@@ -179,13 +162,7 @@ describe('ERC20VotingSetup', function () {
       });
 
       const data = abiCoder.encode(
-        [
-          'uint64',
-          'uint64',
-          'uint64',
-          'tuple(address,string,string)',
-          'tuple(address[],uint256[])',
-        ],
+        prepareInstallDataTypes,
         [
           1,
           1,
@@ -240,13 +217,7 @@ describe('ERC20VotingSetup', function () {
       });
 
       const data = abiCoder.encode(
-        [
-          'uint64',
-          'uint64',
-          'uint64',
-          'tuple(address,string,string)',
-          'tuple(address[],uint256[])',
-        ],
+        prepareInstallDataTypes,
         [
           1,
           1,
@@ -283,7 +254,8 @@ describe('ERC20VotingSetup', function () {
       const governanceERC20 = await GovernanceERC20.deploy(
         targetDao.address,
         'name',
-        'symbol'
+        'symbol',
+        { receivers: [], amounts: []}
       );
 
       const nonce = await ethers.provider.getTransactionCount(
@@ -296,13 +268,7 @@ describe('ERC20VotingSetup', function () {
       });
 
       const data = abiCoder.encode(
-        [
-          'uint64',
-          'uint64',
-          'uint64',
-          'tuple(address,string,string)',
-          'tuple(address[],uint256[])',
-        ],
+        prepareInstallDataTypes,
         [1, 1, 1, [governanceERC20.address, '', ''], [[], []]]
       );
 
@@ -349,13 +315,10 @@ describe('ERC20VotingSetup', function () {
         from: erc20VotingSetup.address,
         nonce: nonce,
       });
-      const anticipatedMerkleMinterAddress = ethers.utils.getContractAddress({
-        from: erc20VotingSetup.address,
-        nonce: nonce + 1,
-      });
+      
       const anticipatedPluginAddress = ethers.utils.getContractAddress({
         from: erc20VotingSetup.address,
-        nonce: nonce + 2,
+        nonce: nonce + 1,
       });
 
       const {plugin, helpers, permissions} =
@@ -365,12 +328,11 @@ describe('ERC20VotingSetup', function () {
         );
 
       expect(plugin).to.be.equal(anticipatedPluginAddress);
-      expect(helpers.length).to.be.equal(2);
+      expect(helpers.length).to.be.equal(1);
       expect(helpers).to.be.deep.equal([
-        anticipatedTokenAddress,
-        anticipatedMerkleMinterAddress,
+        anticipatedTokenAddress
       ]);
-      expect(permissions.length).to.be.equal(7);
+      expect(permissions.length).to.be.equal(4);
       expect(permissions).to.deep.equal([
         [
           Op.Grant,
@@ -399,28 +361,7 @@ describe('ERC20VotingSetup', function () {
           targetDao.address,
           AddressZero,
           MINT_PERMISSION_ID,
-        ],
-        [
-          Op.Grant,
-          anticipatedTokenAddress,
-          anticipatedMerkleMinterAddress,
-          AddressZero,
-          MINT_PERMISSION_ID,
-        ],
-        [
-          Op.Grant,
-          anticipatedMerkleMinterAddress,
-          targetDao.address,
-          AddressZero,
-          MERKLE_MINT_PERMISSION_ID,
-        ],
-        [
-          Op.Grant,
-          anticipatedMerkleMinterAddress,
-          targetDao.address,
-          AddressZero,
-          UPGRADE_PERMISSION_ID,
-        ],
+        ]
       ]);
     });
 
@@ -428,13 +369,7 @@ describe('ERC20VotingSetup', function () {
       const daoAddress = targetDao.address;
 
       const data = abiCoder.encode(
-        [
-          'uint64',
-          'uint64',
-          'uint64',
-          'tuple(address,string,string)',
-          'tuple(address[],uint256[])',
-        ],
+        prepareInstallDataTypes,
         [
           participationRequiredPct,
           supportRequiredPct,
@@ -451,13 +386,9 @@ describe('ERC20VotingSetup', function () {
         from: erc20VotingSetup.address,
         nonce: nonce,
       });
-      const anticipatedMerkleMinterAddress = ethers.utils.getContractAddress({
-        from: erc20VotingSetup.address,
-        nonce: nonce + 1,
-      });
       const anticipatedPluginAddress = ethers.utils.getContractAddress({
         from: erc20VotingSetup.address,
-        nonce: nonce + 2,
+        nonce: nonce + 1,
       });
 
       await erc20VotingSetup.prepareInstallation(daoAddress, data);
@@ -491,21 +422,6 @@ describe('ERC20VotingSetup', function () {
       expect(await governanceTokenContract.getDAO()).to.be.equal(daoAddress);
       expect(await governanceTokenContract.name()).to.be.equal(tokenName);
       expect(await governanceTokenContract.symbol()).to.be.equal(tokenSymbol);
-
-      const MerkleMinterFactory = await ethers.getContractFactory(
-        'MerkleMinter'
-      );
-      const merkleMinterContract = MerkleMinterFactory.attach(
-        anticipatedMerkleMinterAddress
-      );
-
-      expect(await merkleMinterContract.getDAO()).to.be.equal(daoAddress);
-      expect(await merkleMinterContract.token()).to.be.equal(
-        anticipatedTokenAddress
-      );
-      expect(await merkleMinterContract.distributorBase()).to.be.equal(
-        await erc20VotingSetup.distributorBase()
-      );
     });
   });
 
@@ -543,29 +459,31 @@ describe('ERC20VotingSetup', function () {
 
     it('correctly returns permissions, when the required number of helpers is supplied', async () => {
       const plugin = ethers.Wallet.createRandom().address;
-      const GovernanceTokenFactory = await ethers.getContractFactory(
+      const GovernanceERC20 = await ethers.getContractFactory(
         'GovernanceERC20'
       );
-      const token = await GovernanceTokenFactory.deploy(
+      const GovernanceWrappedERC20 = await ethers.getContractFactory(
+        'GovernanceWrappedERC20'
+      );
+      const governanceERC20 = await GovernanceERC20.deploy(
         targetDao.address,
+        tokenName,
+        tokenSymbol,
+        {receivers: [], amounts:[]}
+      );
+
+      const governanceWrappedERC20 = await GovernanceWrappedERC20.deploy(
+        governanceERC20.address,
         tokenName,
         tokenSymbol
       );
-      const MerkleMinterFactory = await ethers.getContractFactory(
-        'MerkleMinter'
-      );
-      const merkleMinter = await MerkleMinterFactory.deploy();
-      await merkleMinter.initialize(
-        targetDao.address,
-        token.address,
-        AddressZero
-      );
-
+      
+      // When the helpers contain governanceWrappedERC20 token
       const permissions1 =
         await erc20VotingSetup.callStatic.prepareUninstallation(
           targetDao.address,
           plugin,
-          [token.address],
+          [governanceWrappedERC20.address],
           EMPTY_DATA
         );
 
@@ -600,40 +518,19 @@ describe('ERC20VotingSetup', function () {
         await erc20VotingSetup.callStatic.prepareUninstallation(
           targetDao.address,
           plugin,
-          [token.address, merkleMinter.address],
+          [governanceERC20.address],
           EMPTY_DATA
         );
 
-      expect(permissions2.length).to.be.equal(7);
+      expect(permissions2.length).to.be.equal(4);
       expect(permissions2).to.deep.equal([
         ...essentialPermissions,
         [
           Op.Revoke,
-          token.address,
+          governanceERC20.address,
           targetDao.address,
           AddressZero,
           MINT_PERMISSION_ID,
-        ],
-        [
-          Op.Revoke,
-          token.address,
-          merkleMinter.address,
-          AddressZero,
-          MINT_PERMISSION_ID,
-        ],
-        [
-          Op.Revoke,
-          merkleMinter.address,
-          targetDao.address,
-          AddressZero,
-          MERKLE_MINT_PERMISSION_ID,
-        ],
-        [
-          Op.Revoke,
-          merkleMinter.address,
-          targetDao.address,
-          AddressZero,
-          UPGRADE_PERMISSION_ID,
         ],
       ]);
     });