+++ packages/contracts/CHANGELOG.md
@@ -39,6 +39,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added an abstract `ERC165RegistryBase` `Component` to register contracts by their address based on their ERC165 interface ID.
 - Added a concrete `ERC165Registry` implementation.
 - Added ENS support for `PluginRepoRegistry`.
+- Added minting functionality to the `initialize` function of `GovernanceERC20`.
 
 ### Changed
 
@@ -83,6 +84,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Removed `deepEqual` overwrite of `equal` property in Chai Assertion used for testing of emitted events.
 - Removed `ERC165Registry`.
 - Removed `Component` and `MetaTxComponent`.
+- Removed `MerkleMinter` deployment from `ERC20VotingSetup`.
 
 ## v0.2.0-alpha